Add documentation for `CommandHandler` and `EventHandler`.
